requirements.txt
apify==1.7.2
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
